connect to backend over TLS (no certificate verification for now)
adding certificate verification means we need a certificate chain store
somewhere, that is shared between HTTP, HTTPS and TCP proxies,
and for which we need a way to add or remove certificates depending on
the application (so, probably new messages)

There is also a bug with the epoll events after the first handshake:
- since we expect to write first on the backend connection, only the
writable interest is set up
- parts of the handshake will happen when receiving data from the backend
- right now when in "interrupted handshake", we return Continue instead
of WouldBlock, so the proxy will retry again and again. But that means
it busy loops while waiting for the backend to answer. Instead, it
should return WouldBlock, and indicate it wants to read from the
backend.
But I don't know if setting up the "readable" event on the backend
connection could break things for other protocols