fix performance of state hashing
with the change in routing configuration, during hashing the code was going
through the entire list of frontends, and this for each application.
Now each frontend will be seen once. At the same time the frontnds are
now stored in a BTreeMap, to guarantee that we will always see them in
the same order (otherwise the resulting hash can change)